Moved LoadStylesheet to GUIHelpers
... though decided against updating it on the fly in settings.
